DAG: fixing Directed Acyclic scene Graph traversal while the graph structure is modified (adding/deleting nodes) during the traversal.
The traversal flags per visitor are now stored as a map on DAGNode* and the scene graph structure (that can change) is now used for the traversal. Rather than performing the traversal on a precomputed, static sub-graph structure that was not considering the scene graph structure changes.

Note: there is still an undetermined case, when a visitor is run from a sub-graph that contains a multi-node linked with a node outside of the sub-graph.

git-svn-id: svn://scm.gforge.inria.fr/svn/sofa/trunk/Sofa@9721 b7f1a160-662a-0410-a6d8-bea165342f16